Start having strong opinions on autogeneration
It's important to not be too aspirational about what we can accomplish
with autogeneration. Specifically, the massive switch statement is a
large challenge because not all APIs are consistent. There are some
ways around this, but let's solve those later.

We also need to autogenerate command line flags for each subcommand.
In order to this more easily with a sane syntax, I switched from hcl
to toml, which is a little more lightweight and a better fit fro this
situation.