Add EditDNSRecord.
This presents a small foot-gun that I'm mostly unhappy with.

When Editing a DNS record, one must remember to set every detail
of the record each time, otherwise the DNS record will revert back
to it's default value.

Forgetting a `--proxy` means that your origin will no longer be
proxied.